Implement z-index, if-then, and grouping to work for gifs for nft-chef (#19)
* Test and add z-index, grouping, if-then

* Update and make rarity work

* Update readme

* Update readme

* Update readme

+++ utils/file.py
@@ -13,10 +13,19 @@ def parse_global_config() -> dict:
     return global_config_json
 
 
-def setup_directory(directory_path: str) -> None:
-    if os.path.exists(directory_path):
+global_config_json = parse_global_config()
+is_debug = global_config_json["debug"]
+
+
+def setup_directory(directory_path: str, delete_if_exists: bool=True) -> None:
+    is_debug and print(f"Setting up directory {directory_path}")
+    if os.path.exists(directory_path) and delete_if_exists:
+        is_debug and print(f"Directory {directory_path} exists, deleting directory")
         shutil.rmtree(directory_path)
-    os.makedirs(directory_path)
+
+    if not os.path.exists(directory_path):
+        is_debug and print(f"Creating directory {directory_path}")
+        os.makedirs(directory_path)
 
 
 def get_png_file_name(file_name: str) -> str:
